requirements.txt
asgiref==3.7.2
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.3
django-taggit==4.0.0
djangorestframework==3.14.0
Faker==19.3.1
Markdown==3.4.4
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
